.github/scripts/requirements.txt
PyGithub==2.5.0
PyYAML==6.0.2
requests==2.32.3
GitPython==3.1.43
black==24.10.0
